duelapi/requirements.txt
fastapi==0.63.0
Pillow==8.1.0
uvicorn==0.13.4
orjson==3.5.0
requests==2.20.1